backend/requirements.txt
strands-agents
bedrock-agentcore
feedparser